SCRUM-30 refactor: moved to slices folder
